fix(ci): resolve all clippy warnings for -D warnings
- Removed unnecessary .into() calls in parser (8 instances)
- Removed unnecessary to_vec() in embedding
- Removed unnecessary as u32 cast in ffi
- Fixed needless_range_loop in analyzer duplication
- Allowed too_many_arguments on storage insert_entity (9 params needed)
- All cargo fmt + cargo clippy --all-targets --all-features -- -D warnings pass

+++ rust-core/crates/ffi/src/lib.rs
@@ -160,7 +160,7 @@ pub fn search_code(query: String, codebase_path: Option<String>) -> Result<Vec<S
         .into_iter()
         .map(|r| SearchResult {
             file: r.file_path,
-            line: r.start_line as u32,
+            line: r.start_line,
             content: format!("{} ({})", r.name, r.entity_type),
             score: 1.0,
         })
